+++ handler/security/security_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/burgesQ/gommon/webtest"
 	"github.com/burgesQ/webfmwk/v5"
+	"github.com/stretchr/testify/require"
 )
 
 const (
@@ -14,10 +15,12 @@ const (
 )
 
 func TestHandler(t *testing.T) {
-	s := webfmwk.InitServer(webfmwk.CheckIsUp(),
+	s, e := webfmwk.InitServer(webfmwk.CheckIsUp(),
 		webfmwk.SetPrefix("/api"),
 		webfmwk.WithHandlers(Handler),
 	)
+
+	require.Nil(t, e)
 
 	t.Cleanup(func() {
 		ctx := s.GetContext()